Give rounding intrinsics their own modules
