+++ crates/core_simd/src/round.rs
@@ -2,12 +2,12 @@ macro_rules! implement {
     {
         $type:ident, $int_type:ident
     } => {
+        #[cfg(feature = "std")]
         impl<const LANES: usize> crate::$type<LANES>
         where
             Self: crate::LanesAtMost32,
         {
             /// Returns the largest integer less than or equal to each lane.
-            #[cfg(feature = "std")]
             #[must_use = "method returns a new vector and does not mutate the original value"]
             #[inline]
             pub fn floor(self) -> Self {
@@ -15,7 +15,6 @@ macro_rules! implement {
             }
 
             /// Returns the smallest integer greater than or equal to each lane.
-            #[cfg(feature = "std")]
             #[must_use = "method returns a new vector and does not mutate the original value"]
             #[inline]
             pub fn ceil(self) -> Self {
